requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic-settings==2.10.1
python-dotenv==1.1.1
python-multipart==0.0.20
